distributemesh doesnt return in single processor runs
removed the check and return from the top of distributemesh when running single
process jobs.
This check and behaviour is the reasonable thing to do, however the
communicators are also initialized in this function (even though they
are useles).
The only added check is in calling metis: if we are running a single
process jog we skip calling metis as it seg faults when asked to
partition into one group.
This is not worth looking into any further as in real world situation
one would use pure mmg3d for single process jobs