avoid sum of memMax exceeding memGloMax in merge_grps
in the loop where internal points, interface tetra and internal tetra
are merged to a single mesh we were updateing the mesh0->memMax and
freeing the meshes that we were merging. This resulted in mesh0->memMax
exceeding the memGloMax limit and any allocations happening for mesh0
while mesh was being merged to mesh0 could result in exceeding the
globally allowed memory limit.

the fix is to move updateing mesh0->memMax and freeing the rest of the
meshes in a loop immidiately after the merge loop.