fix tests failing on plafrim: type conversion errors in allocations
the tests on plafrim were failing with an out of memory error. That
was very weird because it was happening in the beginning when resizing
buffers to use all the memory authorized by the user.

some weird out of bounds buffer accesses that the address sanitizer
complained about indicated where this was happening and by inspecting
the code and enabling more compiler options (-Wall and -Wextra) revealed
that the issue was happening because of floats passed as newsize.

In this commit I tried to bulletproof the allocation macros and also
appears to fix the tests on plafrim.