Remove useless code in PMMG_updateTag function
When a required edge is added to the hash table, it is added without any tag, then we add the MG_BDY tag so the first time that we call hGet, we found gettag==MG_BDY and the test gettag is not MG_REQ answers always. In consequence, we add the MG_REQ + MG_NOSURF tag  to all the edges (even the user required once).

Note that user required edges are nevertheless preserved because during the step 4 we check the tag that is stored in each xtetra (and user's required tags have been preserved) and we remove the spurious NOSURF tag from the tag that is founded in the hash table.