Remove ununsed function
In a first place, Luca, that implemented the // analysis, guessed that some part of the analysis (such as the analysis of non-manifold situations)  were not
possible at the beginning of the process and have to be updated after interface migration. Finally, the non-manifold analysis is done during the pre-processing step and the 'update_nmgeom' function is not called anymore and can be removed.